Add tests and fix sync executor retry logic
