+++ tests/experimental/evals/functions/test_executor.py
@@ -1,0 +1,193 @@
+import asyncio
+import signal
+import nest_asyncio
+
+from phoenix.experimental.evals.functions.executor import (
+    AsyncExecutor,
+    SyncExecutor,
+    get_executor_on_sync_context,
+)
+
+# AsyncExecutor tests
+
+
+async def test_async_executor_executes():
+    async def dummy_fn(payload: int) -> int:
+        return payload - 1
+
+    executor = AsyncExecutor(dummy_fn, concurrency=10, max_retries=0)
+    inputs = [1, 2, 3, 4, 5]
+    outputs = await executor.execute(inputs)
+    assert outputs == [0, 1, 2, 3, 4]
+
+
+async def test_async_executor_executes_many_tasks():
+    async def dummy_fn(payload: int) -> int:
+        return payload
+
+    executor = AsyncExecutor(dummy_fn, concurrency=10, max_retries=0)
+    inputs = [x for x in range(100)]
+    outputs = await executor.execute(inputs)
+    assert outputs == inputs
+
+
+def test_async_executor_runs_synchronously():
+    async def dummy_fn(payload: int) -> int:
+        return payload - 2
+
+    executor = AsyncExecutor(dummy_fn, concurrency=10, max_retries=0)
+    inputs = [1, 2, 3, 4, 5]
+    outputs = executor.run(inputs)
+    assert outputs == [-1, 0, 1, 2, 3]
+
+
+async def test_async_executor_execute_exits_early_on_error():
+    async def dummy_fn(payload: int) -> int:
+        if payload == 3:
+            raise ValueError("test error")
+        return payload - 1
+
+    executor = AsyncExecutor(
+        dummy_fn, concurrency=1, max_retries=0, exit_on_error=True, fallback_return_value=52
+    )
+    inputs = [1, 2, 3, 4, 5]
+    outputs = await executor.execute(inputs)
+    assert outputs == [0, 1, 52, 52, 52]
+
+
+def test_async_executor_run_exits_early_on_error():
+    async def dummy_fn(payload: int) -> int:
+        if payload == 3:
+            raise ValueError("test error")
+        return payload - 1
+
+    executor = AsyncExecutor(
+        dummy_fn, concurrency=1, max_retries=0, exit_on_error=True, fallback_return_value=52
+    )
+    inputs = [1, 2, 3, 4, 5]
+    outputs = executor.run(inputs)
+    assert outputs == [0, 1, 52, 52, 52]
+
+
+async def test_async_executor_can_continue_on_error():
+    async def dummy_fn(payload: int) -> int:
+        if payload == 3:
+            raise ValueError("test error")
+        return payload - 1
+
+    executor = AsyncExecutor(
+        dummy_fn, concurrency=1, max_retries=0, exit_on_error=False, fallback_return_value=52
+    )
+    inputs = [1, 2, 3, 4, 5]
+    outputs = await executor.execute(inputs)
+    assert outputs == [0, 1, 52, 3, 4]
+
+
+async def test_async_executor_sigint_handling():
+    async def async_fn(x):
+        await asyncio.sleep(0.01)
+        return x
+
+    executor = AsyncExecutor(async_fn, concurrency=5, max_retries=0, fallback_return_value="test")
+
+    # Run the executor with a large number of inputs
+    task = asyncio.create_task(executor.execute(list(range(100))))
+    await asyncio.sleep(0.1)
+
+    # Simulate a SIGINT signal
+    executor._signal_handler(signal.SIGINT, None)
+    results = await task
+
+    assert len(results) == 100
+    assert results.count("test") > 0, "some inputs should not have been processed"
+
+
+# SyncExecutor tests
+
+
+def test_sync_executor_runs_many_tasks():
+    def dummy_fn(payload: int) -> int:
+        return payload
+
+    executor = SyncExecutor(dummy_fn, max_retries=0)
+    inputs = [x for x in range(1000)]
+    outputs = executor.run(inputs)
+    assert outputs == inputs
+
+
+def test_sync_executor_runs():
+    def dummy_fn(payload: int) -> int:
+        return payload - 2
+
+    executor = SyncExecutor(dummy_fn, max_retries=0)
+    inputs = [1, 2, 3, 4, 5]
+    outputs = executor.run(inputs)
+    assert outputs == [-1, 0, 1, 2, 3]
+
+
+def test_sync_executor_run_exits_early_on_error():
+    def dummy_fn(payload: int) -> int:
+        if payload == 3:
+            raise ValueError("test error")
+        return payload - 1
+
+    executor = SyncExecutor(dummy_fn, exit_on_error=True, fallback_return_value=52, max_retries=0)
+    inputs = [1, 2, 3, 4, 5]
+    outputs = executor.run(inputs)
+    assert outputs == [0, 1, 52, 52, 52]
+
+
+def test_sync_executor_can_continue_on_error():
+    def dummy_fn(payload: int) -> int:
+        if payload == 3:
+            raise ValueError("test error")
+        return payload - 1
+
+    executor = SyncExecutor(dummy_fn, exit_on_error=False, fallback_return_value=52, max_retries=0)
+    inputs = [1, 2, 3, 4, 5]
+    outputs = executor.run(inputs)
+    assert outputs == [0, 1, 52, 3, 4]
+
+
+async def test_executor_factory_returns_sync_in_async_context():
+    def sync_fn():
+        pass
+
+    async def async_fn():
+        pass
+
+    async def executor_in_async_context():
+        return get_executor_on_sync_context(sync_fn, async_fn)
+
+    executor = await executor_in_async_context()
+    assert isinstance(executor, SyncExecutor)
+
+
+async def test_executor_factory_returns_async_in_patched_async_context():
+    nest_asyncio.apply()
+
+    def sync_fn():
+        pass
+
+    async def async_fn():
+        pass
+
+    async def executor_in_async_context():
+        return get_executor_on_sync_context(sync_fn, async_fn)
+
+    executor = await executor_in_async_context()
+    assert isinstance(executor, AsyncExecutor)
+
+
+def test_executor_factory_returns_async_in_sync_context():
+    def sync_fn():
+        pass
+
+    async def async_fn():
+        pass
+
+    def executor_in_sync_context():
+        return get_executor_on_sync_context(sync_fn, async_fn)
+
+    executor = executor_in_sync_context()
+    assert isinstance(executor, AsyncExecutor)
